# sql: avoid unnecessary singleflight requests for role membership cache