--- a/pkg/bench/rttanalysis/testdata/benchmark_expectations
+++ b/pkg/bench/rttanalysis/testdata/benchmark_expectations
@@ -126,6 +126,8 @@ exp,benchmark
 20,Truncate/truncate_2_column_1_rows
 20,Truncate/truncate_2_column_2_rows
 3,UDFResolution/select_from_udf
+9,UseManyRoles/use_2_roles
+3,UseManyRoles/use_50_roles
 1,VirtualTableQueries/select_crdb_internal.invalid_objects_with_1_fk
 1,VirtualTableQueries/select_crdb_internal.tables_with_1_fk
 5,VirtualTableQueries/virtual_table_cache_with_point_lookups