… docs/generated/settings/settings-for-tenants.txt
… docs/generated/settings/settings.html
… pkg/sql/user_test.go