See individual commits for refactors that make the enhancement easier.

---

Since we use a query that scans the whole role_members table to
populate the membership cache, it is wasteful to launch multiple
singleflight requests for different users. This patch changes
the cache refresh logic of the singleflight so that it always populates
the entire cache, which means there will only ever be one in-flight
query per each node.

```
$ benchdiff --old 98b0521cb18 --new 26daef27780 . -r BenchmarkUseManyRoles

pkg=1/1 iter=10/10 cockroachdb/cockroach/pkg/bench/rttanalysis \

name                          old time/op     new time/op     delta
UseManyRoles/use_50_roles-10      225ms ± 6%      114ms ± 2%  -49.43%  (p=0.000 n=10+10)
UseManyRoles/use_2_roles-10      10.7ms ± 4%      7.4ms ± 3%  -30.74%  (p=0.000 n=10+8)

name                          old roundtrips  new roundtrips  delta
UseManyRoles/use_2_roles-10        9.00 ± 0%       5.00 ± 0%  -44.44%  (p=0.000 n=10+10)
UseManyRoles/use_50_roles-10       3.00 ± 0%       3.00 ± 0%     ~     (all equal)

name                          old alloc/op    new alloc/op    delta
UseManyRoles/use_50_roles-10      438MB ± 0%      377MB ± 1%  -13.88%  (p=0.000 n=9+10)
UseManyRoles/use_2_roles-10      20.7MB ± 2%     19.1MB ± 1%   -7.46%  (p=0.000 n=10+10)

name                          old allocs/op   new allocs/op   delta
UseManyRoles/use_50_roles-10       881k ± 0%       398k ± 4%  -54.81%  (p=0.000 n=8+10)
UseManyRoles/use_2_roles-10       54.4k ± 8%      36.0k ± 5%  -33.85%  (p=0.000 n=10+10)
```

fixes https://github.com/cockroachdb/cockroach/issues/135931
Release note (performance improvement): Improved the internal caching
logic for role membership information. This reduces the latency impact
of commands such as `DROP ROLE`, `CREATE ROLE`, and `GRANT role TO user`,
which cause the role membership cache to be invalidated.