@rafiss 2024-11-23
I created a simple read-only workload that runs concurrent queries with a thousand different users, and ran it on a 3 node cluster.

Before this PR, CREATE ROLE and DROP ROLE caused a spike on the 99.9th percentile latency and also the rate of internal queries.
```
root@localhost:26257/defaultdb> select timestamp, "eventType" from system.eventlog where timestamp > '2024-11-22 23:07'::timestamptz and timestamp < '2024-11-22 23:11'::timestamptz
                             -> order by timestamp asc;
          timestamp          |  eventType
-----------------------------+--------------
  2024-11-22 23:08:57.942819 | create_role
  2024-11-22 23:09:27.021219 | drop_role
```
<img width="906" alt="image" src="https://github.com/user-attachments/assets/82e82590-ef57-43d4-abe4-4e28c9caa086">
<img width="912" alt="image" src="https://github.com/user-attachments/assets/268cdb16-be3c-4e25-8f5a-841ff0e153e3">


With this PR, there was no noticeable impact.
```
root@localhost:26257/defaultdb> select timestamp, "eventType" from system.eventlog where timestamp > '2024-11-22 23:53'::timestamptz and timestamp < '2024-11-22 23:57'::timestamptz
                             -> order by timestamp asc;
          timestamp          |  eventType
-----------------------------+--------------
  2024-11-22 23:54:47.286715 | create_role
  2024-11-22 23:55:29.026177 | drop_role
```
<img width="901" alt="image" src="https://github.com/user-attachments/assets/8894a176-4363-4595-aaba-96bdb2d00d31">
<img width="903" alt="image" src="https://github.com/user-attachments/assets/7a8f1e3c-30c2-4957-a3c6-fcc1d5aa3c29">


<details>
<summary>Click to see the workload script</summary>
```
package main

import (
	"context"
	"fmt"
	"github.com/jackc/pgx/v5"
	"github.com/jackc/pgx/v5/pgxpool"
	"golang.org/x/sync/errgroup"
	"math/rand"
	"strconv"
	"strings"
)

func main() {
	u := "<node 1 connection string redacted>"
	ctx := context.Background()
	conn, err := pgx.Connect(ctx, u)
	if err != nil {
		panic(err)
	}

	fmt.Println("Creating table ...")

	_, err = conn.Exec(ctx, "DROP TABLE IF EXISTS tab")
	if err != nil {
		panic(err)
	}
	_, err = conn.Exec(ctx, "CREATE TABLE IF NOT EXISTS tab (a INT PRIMARY KEY)")
	if err != nil {
		panic(err)
	}
	_, err = conn.Exec(ctx, "INSERT INTO tab VALUES (1)")
	if err != nil {
		panic(err)
	}

	fmt.Println("Creating roles ...")

	_, err = conn.Exec(ctx, "CREATE ROLE IF NOT EXISTS parent_role")
	if err != nil {
		panic(err)
	}

	_, err = conn.Exec(ctx, "GRANT ALL ON TABLE tab TO parent_role")
	if err != nil {
		panic(err)
	}

	roles := make([]string, 0, 1000)
	for i := 0; i < 1000; i++ {
		roleName := "role" + strconv.Itoa(i)
		_, err = conn.Exec(ctx, "CREATE ROLE IF NOT EXISTS "+roleName)
		if err != nil {
			panic(err)
		}
		roles = append(roles, roleName)

		if i%50 == 0 {
			fmt.Println("Created role", i)
		}
	}

	_, err = conn.Exec(ctx, "GRANT parent_role TO "+strings.Join(roles, ", "))
	if err != nil {
		panic(err)
	}

	urls := []string{
		"<node 1 connection string redacted>",
		"<node 2 connection string redacted>",
		"<node 3 connection string redacted>",
	}
	pools := make([]*pgxpool.Pool, len(urls))

	fmt.Println("Creating pools ...")
	for i, u := range urls {
		cfg, err := pgxpool.ParseConfig(u)
		if err != nil {
			panic(err)
		}
		cfg.MaxConns = 500
		pool, err := pgxpool.NewWithConfig(ctx, cfg)
		if err != nil {
			panic(err)
		}
		pools[i] = pool
	}

	g, _ := errgroup.WithContext(ctx)
	g.SetLimit(1000)

	fmt.Println("Sending load ...")

	i := 0
	for {
		pool := pools[rand.Intn(len(pools))]
		roleID := i % 1000
		g.Go(func() error {
			conn, err := pool.Acquire(ctx)
			if err != nil {
				return err
			}
			defer conn.Release()

			_, err = conn.Exec(ctx, "SET ROLE role"+strconv.Itoa(roleID))
			if err != nil {
				return err
			}
			var res int
			err = conn.QueryRow(ctx, "SELECT * FROM tab LIMIT 1", pgx.QueryExecModeSimpleProtocol).Scan(&res)
			if err != nil {
				return err
			}

			return nil
		})
		if i%100 == 0 {
			fmt.Println("Sent load", i)
		}
		i++
	}
}

```
</details>

With larger node counts and greater numbers of concurrent active SQL users, the impact would be greater.